moved study helper functions to utils.py
